# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Server, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "servers": [
    {
      "url": "https://api.revmetrix.io"
    }
  ],
  "paths": {
...
    "/api/posts/StartPractice": {
      "post": {
        "tags": [
          "Posts"
        ],
        "operationId": "StartPractice",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Practice"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Practice"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Practice"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
    "/api/posts/StartSession": {
      "post": {
        "tags": [
          "Posts"
        ],
        "operationId": "StartSession",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Session"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Session"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Session"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "boolean"
                }
              },
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              },
              "text/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Practice": {
        "type": "object",
        "properties": {
          "event_id": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
      "Session": {
        "type": "object",
        "properties": {
          "leauge_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "tournament_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "practice_id": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "time": {
            "type": "string",
            "format": "date-time"
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "total_Games": {
            "type": "integer",
            "format": "int32"
          },
          "total_Frames": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}